requirements.txt
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
Deprecated==1.2.14
google-api-core==2.11.0
google-auth==2.19.0
google-cloud-trace==1.11.1
googleapis-common-protos==1.59.0
grpcio==1.54.2
grpcio-status==1.54.2
idna==3.4
importlib-metadata==6.0.1
opentelemetry-api==1.18.0
opentelemetry-exporter-gcp-trace==1.5.0
opentelemetry-resourcedetector-gcp==1.5.0a0
opentelemetry-sdk==1.18.0
opentelemetry-semantic-conventions==0.39b0
proto-plus==1.22.2
protobuf==4.23.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
requests==2.31.0
rsa==4.9
six==1.16.0
typing_extensions==4.6.2
urllib3==1.26.16
wrapt==1.15.0
zipp==3.15.0
